[Configury] Search for and link against liblua.
I've discovered that the best way to avoid crashiness when running
Lua on Mac OS 10.6 with runtime binary modules is to build lua itself
with a liblua.dylib, and then link the lua interpreter and all the
modules slavishly against that same dylib. This patch makes sure
lcurses behaves accordingly.